refector:챗봇 컴포넌트 분류
